requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
requests>=2.31
pyarrow>=14.0
pandas>=2.1
google-cloud-storage>=2.14
boto3>=1.34
sqlalchemy>=2.0
